Заменил __int64 на int. Выполнил задание.

+++ Lab2_q4/GeneratePrimeNumbersSet/GeneratePrimeNumbersSet/GeneratePrimeNumbersSet.h
@@ -1,3 +1,3 @@
 #pragma once
 #include <set>
-std::set<__int64> GeneratePrimeNumSet(__int64 upBound);
+std::set<int> GeneratePrimeNumSet(int upBound);
